C# file contain a single class at the root level :recycle:
