+++ Src/Notion.Client/Api/Databases/RequestParams/DatabasesCreateParameters/PropertySchema/MultiSelectOptionSchema.cs
@@ -1,0 +1,6 @@
+﻿namespace Notion.Client
+{
+    public class MultiSelectOptionSchema : SelectOptionSchema
+    {
+    }
+}
