+++ Src/Notion.Client/Api/Databases/RequestParams/DatabasesCreateParameters/PropertySchema/MultiSelectPropertySchema.cs
@@ -7,8 +7,4 @@ namespace Notion.Client
         [JsonProperty("multi_select")]
         public OptionWrapper<MultiSelectOptionSchema> MultiSelect { get; set; }
     }
-
-    public class MultiSelectOptionSchema : SelectOptionSchema
-    {
-    }
 }
