+++ Src/Notion.Client/Api/Databases/RequestParams/DatabasesCreateParameters/PropertySchema/SelectPropertySchema.cs
@@ -1,18 +1,7 @@
-﻿using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
-namespace Notion.Client
+﻿namespace Notion.Client
 {
     public class SelectPropertySchema : IPropertySchema
     {
         public OptionWrapper<SelectOptionSchema> Select { get; set; }
-    }
-
-    public class SelectOptionSchema
-    {
-        public string Name { get; set; }
-
-        [JsonConverter(typeof(StringEnumConverter))]
-        public Color Color { get; set; }
     }
 }
