+++ Src/Notion.Client/Api/Databases/RequestParams/DatabasesCreateParameters/RichTextMentionInput.cs
@@ -5,13 +5,4 @@
         public override RichTextType Type => RichTextType.Mention;
         public MentionInput Mention { get; set; }
     }
-
-    public class MentionInput
-    {
-        public string Type { get; set; }
-        public Person User { get; set; }
-        public ObjectId Page { get; set; }
-        public ObjectId Database { get; set; }
-        public DatePropertyValue Date { get; set; }
-    }
 }
